Time: 7 ms (71.82%) | Memory: 8.6 MB (33.96%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        int n = x;
+        long n = x;
         long num = 0;
         if(x<0) return false;
         while(n>0){
@@ -9,6 +9,6 @@ public:
             num = num*10 + rem;
             n /=10;
         }
-        return (int)num==x?true:false;
+        return num==x?true:false;
     }
 };
